services/authentication/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
PyJWT==2.8.0
pyotp==2.9.0
django-qr-code==4.0.1
requests==2.31.0
daphne==4.0.0
channels==4.0.0
psycopg==3.1.18
Twisted[tls,http2]
